requirements.txt
aiogram==3.1.1
python-dotenv==1.0.1
tinkoff-investments==0.2.0b97
